requirements.txt
matplotlib==3.10.6
numba==0.61.2
numpy==2.2.6
pandas==2.3.2
PyYAML==6.0.2
scipy==1.16.2
torch==2.8.0
torchvision==0.23.0
